remove Layout wrapper from AuthForm, rename MobileMenu > MainMenu (disambiguation), move some styles into .style files, move MainMenu into Header, linter fixes, move Type declaration into Header.styles
