+++ components/Header/Header.tsx
@@ -5,57 +5,19 @@ import { HeaderProps } from "./types";
 import Sticky from "react-sticky-el";
 import { useAuth } from "../../config/auth";
 import SearchBar from "../SearchBar";
-import styled from "@emotion/styled";
+import { MainMenu } from "../MainMenu";
+import { menusData } from "../MainMenu/data/menusData";
 
-const TopHeader = styled.div`
-  padding: 30px 0px;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-`;
-const RightSide = styled.div`
-  width: 10%;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-`;
-const LogoDiv = styled.div`
-  padding: 15px 30px;
-  background: grey;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-`;
-const HeaderDiv = styled.header`
-  margin-bottom: 25px;
-`;
-export interface LinkDivProps {
-  isActive: boolean;
-}
-const LinkDiv = styled.a<LinkDivProps>`
-  font-size: 14px;
-  text-decoration: ${(props) => (props.isActive ? "underline" : "none")};
-`;
-const BottomHeader = styled.div`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-  flex-wrap: wrap;
-  padding: 30px 0px;
-  background: black;
-  & > :first-child {
-    padding-left: 0px;
-  }
-  & > :last-child {
-    padding-right: 0px;
-  }
-`;
-const Category = styled.a`
-  padding: 15px;
-`;
+import {
+  TopHeader,
+  RightSide,
+  LogoDiv,
+  HeaderDiv,
+  LinkDiv,
+  BottomHeader,
+  Category
+} from "./Header.styles";
+
 const dummyCategories = ["Best Sellers", "Latest", "Seasonal", "Luxury", "On Sale", "Coming Soon"];
 
 export const Header: React.FC<HeaderProps> = (props) => {
@@ -93,42 +55,14 @@ export const Header: React.FC<HeaderProps> = (props) => {
       </TopHeader>
       <Sticky>
         <BottomHeader>
-          {dummyCategories.map((category) => (
-            <Link href="/" key={category}>
-              <Category>{category}</Category>
-            </Link>
-          ))}
-          <style jsx>{`
-            .bottom-header {
-              display: flex;
-              flex-direction: row;
-              align-items: center;
-              justify-content: space-between;
-              flex-wrap: wrap;
-              padding: 30px 0px;
-              background: black;
-            }
-            .category {
-              padding: 15px;
-            }
-            .bottom-header > :first-child {
-              padding-left: 0px;
-            }
-            .bottom-header > :last-child {
-              padding-right: 0px;
-            }
-
-            header {
-              margin-bottom: 25px;
-            }
-            a {
-              font-size: 14px;
-              text-decoration: none;
-            }
-            .is-active {
-              text-decoration: underline;
-            }
-          `}</style>
+          <MainMenu
+            pcMenuItemClassName={"pc-menu-item"}
+            outterContainerId={"outer-container"}
+            pageWrapId={"page-wrap"}
+            animationType={"slide"}
+            menusData={menusData}
+            right={false}
+          />
           <SearchBar />
         </BottomHeader>
       </Sticky>
