+++ components/Layout/Layout.tsx
@@ -1,69 +1,23 @@
 import React from "react";
-import styled from "@emotion/styled";
 import { ClassNames } from "@emotion/react";
 import { LayoutProps } from "./types";
-import { MobileMenu, Foot } from "../../components";
-import { menusData } from "../MobileMenu/data/menusData";
+import { Header } from "../../components";
+import { Foot } from "../../components";
 import { Column } from "../Foot/Foot";
-const Logo = styled.div`
-  width: 181px;
-  height: 45px;
-  font-size: 24px;
-  line-height: 45px;
-  border: 1px solid #969696;
-  color: #fff;
-  font-weight: bold;
-  font-family: Montserrat;
-  text-align: center;
-  margin-bottom: 60px;
-  @media(max-width:375px){
-    width: 80px;
-    height: 19.83px;
-    font-size: 10px;
-    line-height: 19.83px;
-    margin-bottom: 55.07px;
-  }
-`;
+
+import {
+  Container,
+  Content,
+  CameraIcon,
+  FacebookIcon,
+  PlayIcon,
+  CircleIcon,
+  Logo
+} from "./Layout.styles";
+
 const MyLogo: React.FC = () => <Logo>POL</Logo>;
-const CameraIcon = styled.img`
-  width: 11px;
-  height: auto;
-  margin-right: 3.88px;
-  @media(max-width: 375px){
-    width: 20.71px;
-    height: auto;
-    margin-right: 7.76px;
-  }
-`;
-const FacebookIcon = styled.img`
-  width: 6.81px;
-  height: auto;
-  margin-right: 3.88px;
-  @media(max-width: 375px){
-    width: 13.63px;
-    height: auto;
-    margin-right: 7.76px;
-  }
-`;
-const PlayIcon = styled.img`
-  width: 12.29px;
-  height: auto;
-  margin-right: 3.88px;
-  @media(max-width: 375px){
-    width: 24.59px;
-    height: auto;
-    margin-right: 7.76px;
-  }
-`;
-const CircleIcon = styled.img`
-  width: 10.35px;
-  height: auto;
-  @media(max-width: 375px){
-    width: 20.71px;
-    height: auto;
-  }
-`;
-const iconLinks=[
+
+const iconLinks = [
   {
     icon: <CameraIcon src={"/camera.png"} />,
     url: ""
@@ -80,7 +34,7 @@ const iconLinks=[
     icon: <CircleIcon src={"/circle.png"} />,
     url: ""
   }
-]
+];
 const columns: Column[] = [
   {
     title: "Contact Info",
@@ -89,7 +43,7 @@ const columns: Column[] = [
       "+1 (310) 715-1370",
       "ecom@polclothing.com"
     ],
-    iconLinks,
+    iconLinks
   },
   {
     title: "Information",
@@ -135,31 +89,15 @@ const columns: Column[] = [
     ]
   }
 ];
-export const Container = styled.main`
-  width: 100%;
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  max-height: calc(100% - 235px);
-`;
-export const Content = styled.div`
-  flex: 1;
-  overflow: scroll;
-`;
+
 export const Layout: React.FC<LayoutProps> = ({
   children
 }: {
   children: JSX.Element[] | JSX.Element;
 }) => {
   return (
-    <Container>
-      <MobileMenu
-        pcMenuItemClassName={"pc-menu-item"}
-        outterContainerId={"outter-container"}
-        pageWrapId={"page-wrap"}
-        animationType={"slide"}
-        menusData={menusData}
-        right={false}></MobileMenu>
+    <>
+      <Header />
       <Content>
         {children}
         {/*<Footer
@@ -208,11 +146,11 @@ export const Layout: React.FC<LayoutProps> = ({
                   background-color: #000;
                 `
               }}
-              footerData={{ logo: <MyLogo />, columns,mobileIconLinks:iconLinks }}
+              footerData={{ logo: <MyLogo />, columns, mobileIconLinks: iconLinks }}
             />
           )}
         </ClassNames>
       </Content>
-    </Container>
+    </>
   );
 };
