+++ components/MainMenu/MainMenu.tsx
@@ -8,7 +8,7 @@ import ListItemIcon from "@material-ui/core/ListItemIcon";
 import ExpandLess from "@material-ui/icons/ExpandLess";
 import ExpandMore from "@material-ui/icons/ExpandMore";
 import Collapse from "@material-ui/core/Collapse";
-import { MobileMenuProps, menuDataItem } from "./types";
+import { MainMenuProps, menuDataItem } from "./types";
 import { useMediaQuery } from "react-responsive";
 import DesktopMenu from "./DesktopMenu";
 import styled from "@emotion/styled";
@@ -34,7 +34,7 @@ const MobileHidden = styled.div`
     display: flex;
   }
 `;
-export const MobileMenu = (props: MobileMenuProps) => {
+export const MainMenu = (props: MainMenuProps) => {
   const {
     pcWrapClassName,
     pcMenuItemClassName,
