+++ components/MainMenu/index.tsx
@@ -1,0 +1,1 @@
+export { MainMenu } from "./MainMenu";
