+++ components/MainMenu/snapsvgImporter.ts
