+++ components/MainMenu/types/CrossIcon.d.ts
