+++ components/MainMenu/types/DesktopMenu.d.ts
