+++ components/MainMenu/types/baseStyles.d.ts
