--- components/MobileMenu/index.tsx
@@ -1,1 +1,0 @@
-export { MobileMenu } from "./MobileMenu";
