+++ components/index.ts
@@ -3,7 +3,7 @@ export * from "./Header";
 export * from "./InfoBox";
 export * from "./PostList";
 export * from "./ProductList";
-export * from "./MobileMenu";
+export * from "./MainMenu";
 export * from "./Footer";
 export * from "./AuthForm";
 export * from "./Foot";
