+++ pages/_app.tsx
@@ -4,7 +4,6 @@ import { QueryClient, QueryClientProvider } from "react-query";
 import { Hydrate } from "react-query/hydration";
 import { ReactQueryDevtools } from "react-query/devtools";
 import { AuthProvider } from "../config/auth";
-import { Header } from "../components";
 
 // Styles
 import { ThemeProvider } from "@emotion/react";
@@ -27,7 +26,6 @@ export default function MyApp({ Component, pageProps }: AppProps) {
         <Hydrate state={pageProps.dehydratedState}>
           <ThemeProvider theme={theme}>
             <GlobalStyles />
-            <Header />
             <Component {...pageProps} />
           </ThemeProvider>
         </Hydrate>
